Update Xero crate to use BigDecimal

See merge request !1

+++ src/resources/items.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use client::Client;
 use encoding::{XmlError, XmlSerializable, XmlWriter};
 use error::Error;
@@ -5,7 +6,7 @@ use error::Error;
 #[derive(Debug, Default, Deserialize, Serialize)]
 #[serde(rename_all = "PascalCase")]
 pub struct ItemDetails {
-    pub unit_price: Option<f64>,
+    pub unit_price: Option<BigDecimal>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub account_code: Option<String>,
     #[serde(rename = "COGSAccountCode")]
